Reduce code complexity
Addresses code climate warning
