+++ packages/react-component-library/src/components/TextArea/TextArea.test.tsx
@@ -3,12 +3,12 @@ import '@testing-library/jest-dom/extend-expect'
 import { render, RenderResult } from '@testing-library/react'
 
 import withFormik from '../../enhancers/withFormik'
-import InputProps from '../../types/InputProps'
+import FieldProps from '../../types/FieldProps'
 import FormProps from '../../types/FormProps'
 import { TextArea } from './index'
 
 describe('TextArea', () => {
-  let field: InputProps
+  let field: FieldProps
   let form: FormProps
   let label = ''
   let textInput: RenderResult
@@ -17,7 +17,8 @@ describe('TextArea', () => {
     field = {
       name: 'colour',
       value: '',
-      onChange: jest.fn(),
+      onBlur: null,
+      onChange: jest.fn()
     }
 
     form = {
